@@ -1,23 +1,20 @@
-import Link from "@material-ui/core/Link";
-import makeStyles from "@material-ui/core/styles/makeStyles";
-import Alert from "@material-ui/lab/Alert";
+import Alert from "@mui/material/Alert";
+import Link from "@mui/material/Link";
+import { styled } from "@mui/material/styles";
 import PropTypes from "prop-types";
 import React from "react";
 
-const useStyles = makeStyles((theme) => ({
-  root: {
-    marginTop: theme.spacing(2),
-  },
+const StyledAlert = styled(Alert)(({ theme }) => ({
+  marginTop: theme.spacing(2),
 }));
 
 /** Renders the rights information defined in the used manifest */
 const RightsInformation = ({ rights, t }) => {
-  const { root } = useStyles();
   if (!rights.length) {
     return null;
   }
   return (
-    <Alert className={root} severity="warning">
+    <StyledAlert severity="warning">
       <span>{t("imageCropper.noteRights", { count: rights.length })}: </span>
       {rights.length === 1 ? (
         <Link href={rights[0]} rel="noopener" target="_blank">
@@ -34,7 +31,7 @@ const RightsInformation = ({ rights, t }) => {
           ))}
         </ul>
       )}
-    </Alert>
+    </StyledAlert>
   );
 };
 
