@@ -1,12 +1,12 @@
-import FacebookIcon from "@material-ui/icons/Facebook";
-import MailIcon from "@material-ui/icons/Mail";
-import PinterestIcon from "@material-ui/icons/Pinterest";
-import WhatsAppIcon from "@material-ui/icons/WhatsApp";
+import FacebookIcon from "@mui/icons-material/Facebook";
+import MailIcon from "@mui/icons-material/Mail";
+import PinterestIcon from "@mui/icons-material/Pinterest";
+import WhatsAppIcon from "@mui/icons-material/WhatsApp";
+import X from "@mui/icons-material/X";
 import { MiradorMenuButton } from "mirador/dist/es/src/components/MiradorMenuButton";
 import PropTypes from "prop-types";
 import React from "react";
 
-import XIcon from "../icons/XIcon";
 import { getShareLink } from "../utils";
 
 const iconMapping = {
@@ -14,7 +14,7 @@ const iconMapping = {
   facebook: FacebookIcon,
   pinterest: PinterestIcon,
   whatsapp: WhatsAppIcon,
-  x: XIcon,
+  x: X,
 };
 
 /** Renders a button for sharing the given content on one of the supported providers */
